Bump cargo to 0.1.1.
MINOR BREAKAGE:  Behaviour change - info is now more important than
debug, to align with the log crate from the main rust project.

ADDITIONS: Support added for use as a backend to the log crate.
See the logcrate_use binary for a small example.

To support this, we also now allow a default logger to be set.
Messages for loggers that do now exist will be sent to the default
logger if it is present.